pyapi/piggybank/requirements.txt
ccxt==4.1.43
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pymysql==1.1.0